+++ src/lib/utils/index.ts
@@ -1,26 +1,11 @@
-import { Position, Tile } from "../models/Tile";
+import { Position } from "../models/Tile";
 
 let incrementNum = 0;
 export const getUniqKey = () => incrementNum++;
 
-const getAllPositions = () =>
-  Position.dimensionX.flatMap((x) =>
-    Position.dimensionY.map((y) => ({ x, y })),
-  );
-
 export const getRestPositions = (positions: Position[]): Position[] => {
-  return getAllPositions().filter((ap) =>
+  return Position.all().filter((ap) =>
     positions.every((p) => ap.x !== p.x || ap.y !== p.y),
-  );
-};
-
-export const getTileMap = (tiles: Tile[]) => {
-  return Position.dimensionX.map((x) =>
-    Position.dimensionY.map((y) => {
-      return (
-        tiles.find((t) => t.position.x === x && t.position.y === y) ?? null
-      );
-    }),
   );
 };
 
@@ -31,3 +16,5 @@ export const pick = <T>(candidates: T[], needs: number): T[] => {
   const rest = [...candidates.slice(0, index), ...candidates.slice(index + 1)];
   return [candidates[index], ...pick(rest, needs - 1)];
 };
+
+export * from "./simulate";
